requirements.txt
psycopg2-binary>=2.9.10,<3.0.0
pandas>=2.2.0,<3.0.0
